test(vsock-proto): reorganize exec operation protocol tests
